// src/type/type.ts
export type Personaje = {
    id: number
    name: string
    status: string
    species: string
    type: string
    gender: string
    image: string
    url: string
    created: string
    episode: string[]
    origin: {
        name: string
        url: string
    }
    location: {
        name: string
        url: string
    }

}